Check system container label value
When deciding to setup resolv.conf, skip if the label
io.rancer.container.system has the value 'rancher-agent', not simply if
the label exists.

Making this change because we are going to start adding
io.rancher.container.system=true to all containers that are part of a
system stack so that we can do scheduling based off the presence of
that label.