Reduce network retries from 60 to 15
Motivation: the go-agent that reports a container's IP back to cattle
only waits for 1 minute. So, retrying for 2 minutes is useless and
since the error state is not written until all retries are done,
go-agent doesn't get the more useful error state if this retry period
is longer than go-agent's wait period.